backend/requirements.txt
fastapi==0.115.0
SQLAlchemy==2.0.35
uvicorn==0.30.6
pydantic==2.9.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1